feat: clean up response type enum folders
