+++ packages/core/src/messages/responses/enums/cache/dictionary/index.ts
@@ -1,0 +1,17 @@
+export enum DictionaryGetFieldsResponse {
+  Hit = 'Hit',
+  Miss = 'Miss',
+  Error = 'Error',
+}
+
+export enum DictionaryGetFieldResponse {
+  Hit = 'Hit',
+  Miss = 'Miss',
+  Error = 'Error',
+}
+
+export enum DictionaryFetchResponse {
+  Hit = 'Hit',
+  Miss = 'Miss',
+  Error = 'Error',
+}
