+++ packages/core/src/messages/responses/enums/cache/scalar/index.ts
@@ -1,0 +1,5 @@
+export enum CacheGetResponse {
+  Hit = 'Hit',
+  Miss = 'Miss',
+  Error = 'Error',
+}
