+++ packages/core/src/messages/responses/enums/index.ts
@@ -3,26 +3,4 @@ export enum MomentoResponse {
   Error = 'Error',
 }
 
-export enum CacheGetResponse {
-  Hit = 'Hit',
-  Miss = 'Miss',
-  Error = 'Error',
-}
-
-export enum DictionaryGetFieldsResponse {
-  Hit = 'Hit',
-  Miss = 'Miss',
-  Error = 'Error',
-}
-
-export enum DictionaryGetFieldResponse {
-  Hit = 'Hit',
-  Miss = 'Miss',
-  Error = 'Error',
-}
-
-export enum DictionaryFetchResponse {
-  Hit = 'Hit',
-  Miss = 'Miss',
-  Error = 'Error',
-}
+export * from './cache';
